+++ frontend/.env
@@ -1,0 +1,1 @@
+VITE_URL_HOST_API="http://localhost:8080"
