+++ frontend/src/routes/LogIn.tsx
@@ -1,85 +1,18 @@
 // @ts-nocheck
-import { useContext, useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
 import "./LogIn.css";
-import { UserLoggedContext } from "../contexts/UserLoggedContext";
 import logo from "../assets/images/logoLetras.png";
+import { useLogin } from "../hooks/useLogin";
 
 const LogIn = () => {
-  const { userLogged, setUserLogged } = useContext(UserLoggedContext);
-  const [errors, setErrors] = useState([]);
-  const [user, setUser] = useState({
-    usuario: "",
-    password: "",
-  });
-
-  const [authError, setAuthError] = useState("");
-  const [responseError, setResponseError] = useState(false);
-  const navigate = useNavigate();
-
-  const validateLoginForm = () => {
-    const errors = [];
-
-    if (user.usuario.trim() === "") {
-      errors.push("El correo electrónico es obligatorio");
-    } else if (!/\S+@\S+\.\S+/.test(user.usuario)) {
-      errors.push("El correo electrónico es inválido");
-    }
-
-    if (user.password.trim() === "") {
-      errors.push("No olvides ingresar tu contraseña");
-    } else if (user.password.length < 4 || user.password.length > 15) {
-      errors.push("La contraseña debe contener entre 4 y 15 caracteres");
-    }
-    setErrors(errors);
-    return errors.length === 0;
-  };
-
-  const url = "http://localhost:8080/login";
-  const settings = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(user),
-  };
-
-  const handleLogin = (e) => {
-    e.preventDefault();
-
-    //como no quiero que la persona regrese despues de pasar por el login, lo pongo en true para reemplazar el historial
-    navigate({ replace: true });
-
-    if (validateLoginForm())
-      fetch(url, settings)
-      .then((response) => {
-        if (response.status === 202)
-          response.json()
-          .then(({id, usuario}) => {
-            setUserLogged({
-              id,
-              email: usuario,
-            });
-            navigate("/home");
-          });
-        else {
-          setResponseError(true);
-          setAuthError((response.status === 403) ? "Usuario inactivo" : "Usuario o contraseña no válido");
-        }
-      });
-  };
-
-  const handleMailChange = (e) => {
-    setResponseError(false);
-    setErrors([]);
-    setUser({ ...user, usuario: e.target.value });
-  };
-
-  const handlePasswordChange = (e) => {
-    setResponseError(false);
-    setErrors([]);
-    setUser({ ...user, password: e.target.value });
-  };
+  const {
+    handleLogin,
+    user,
+    errors,
+    responseError,
+    handleMailChange,
+    handlePasswordChange,
+    authError,
+  } = useLogin();
 
   return (
     <section>
